% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosace.R
\name{AddAssayData}
\alias{AddAssayData}
\title{Add Assay Data to the Rosace Object}
\usage{
AddAssayData(object, assay, var.data)
}
\arguments{
\item{object}{A Rosace object}

\item{assay}{An Assay object}

\item{var.data}{Variant's meta data for the Assay object}
}
\value{
The Rosace object with added Assay
}
\description{
Add Assay Data to the Rosace Object
}
