% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosace.R
\name{AddAssaySetData}
\alias{AddAssaySetData}
\title{Add Assayset Object to the Rosace Object}
\usage{
AddAssaySetData(object, assayset)
}
\arguments{
\item{object}{A Rosace object}

\item{assayset}{A AssaySet object}
}
\value{
The Rosace object with added AssaySet
}
\description{
Add Assayset Object to the Rosace Object
}
