% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosette.R
\name{AddDisp}
\alias{AddDisp}
\title{Add dispersion to a Rosette object}
\usage{
AddDisp(object, disp)
}
\arguments{
\item{object}{A Rosette object}

\item{disp}{A numeric value}
}
\value{
A Rosette object with dispersion added
}
\description{
Add dispersion to a Rosette object
}
