% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosace.R
\name{AddScoreData}
\alias{AddScoreData}
\title{Add Score Object to the Rosace Object}
\usage{
AddScoreData(object, score)
}
\arguments{
\item{object}{A Rosace object}

\item{score}{A Score object}
}
\value{
The Rosace object with added Score
}
\description{
Add Score Object to the Rosace Object
}
