% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosace.R
\name{ApplyAssayKey}
\alias{ApplyAssayKey}
\title{Apply function to all Assay objects with the same key}
\usage{
ApplyAssayKey(object, key, fxn, ...)
}
\arguments{
\item{object}{A Rosace object}

\item{key}{The targeted key of Assay object}

\item{fxn}{function to be applied to all Assay objects with the same key}

\item{...}{Additional arguments to be passed to the function}
}
\value{
Rosace object with updated Assay objects
}
\description{
Apply function to all Assay objects with the same key
}
