% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\docType{class}
\name{Assay-class}
\alias{Assay-class}
\alias{Assay}
\alias{AssayGrowth-class}
\alias{AssayGrowth}
\title{The Assay Class}
\description{
The Assay Class

The AssayGrowth Class
}
\section{Slots}{

\describe{
\item{\code{counts}}{A matrix of raw counts
(variants x time points in growth screen)}

\item{\code{var.names}}{A vector of variant names}

\item{\code{key}}{A character string for the experiment}

\item{\code{rep}}{The identity of the replicate}

\item{\code{norm.counts}}{A matrix of normalized counts (variants x time points)}

\item{\code{norm.var.names}}{A vector of variant names for normalized counts}

\item{\code{rounds}}{Number of rounds = Number of time points - 1}
}}

