% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{Assay2AssaySet}
\alias{Assay2AssaySet}
\title{Transform an Assay object to AssaySet object}
\usage{
Assay2AssaySet(assay)
}
\arguments{
\item{assay}{An Assay object}
}
\value{
An AssaySet object
}
\description{
Transform an Assay object to AssaySet object
}
