% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\docType{class}
\name{AssaySet-class}
\alias{AssaySet-class}
\alias{AssaySet}
\alias{AssaySetGrowth-class}
\alias{AssaySetGrowth}
\title{The AssaySet Class}
\description{
The AssaySet Class

The AssaySetGrowth Class
}
\section{Slots}{

\describe{
\item{\code{combined.counts}}{A matrix of counts (variants x cols in reps)}

\item{\code{var.names}}{A vector of variant names corresponding
to the row of combined.counts.}

\item{\code{reps}}{A vector of replicate identity}

\item{\code{key}}{A character string for the experiment}

\item{\code{raw.counts}}{A matrix of raw counts (variants x cols in reps)}

\item{\code{rounds}}{A vector of number of rounds}
}}

