% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA1R.data}
\alias{BRCA1R.data}
\title{BRCA1 RING phage screen: variant info and cell count dataframe}
\format{
A dataframe with processed variant information and cell count across 6 time points
and 6 replicates
}
\usage{
data(BRCA1R)
}
\description{
BRCA1 RING phage screen: variant info and cell count dataframe
}
\keyword{datasets}
