% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckDots}
\alias{CheckDots}
\title{Check the use of the dots}
\usage{
CheckDots(..., args = NULL, fxns = NULL)
}
\arguments{
\item{...}{Arguments passed to other methods}

\item{args}{A list of arguments to check for unused arguments}

\item{fxns}{A list of functions to check for unused arguments}
}
\description{
Function to check the use of unused arguments passed to \code{...}; this
function is designed to be called from another function to see if an
argument passed to \code{...} remains unused and alert the user if so.
Adapted from Seurat::CheckDots
}
