% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSACE.R
\name{CheckStanSetup}
\alias{CheckStanSetup}
\title{Check Stan Setup}
\usage{
CheckStanSetup(mc.cores, install = TRUE, cmdstan_ver = "2.35.0")
}
\arguments{
\item{mc.cores}{Number of cores to use for parallel builds}

\item{install}{if TRUE, install cmdstan to the default path; if FALSE, not update cmdstan
if a string, download to the path specified in string}

\item{cmdstan_ver}{string, cmdstan version to download}
}
\value{
None
}
\description{
Check Stan Setup
}
