% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSACE.R
\name{CompileModel}
\alias{CompileModel}
\title{Compile Stan model}
\usage{
CompileModel(file, print = TRUE)
}
\arguments{
\item{file}{Stan model file path}

\item{print}{whether to print model out model}
}
\value{
CmdStanModel object
}
\description{
Compile Stan model
}
