% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{CreateAssayObject}
\alias{CreateAssayObject}
\alias{CreateAssayGrowthObject}
\title{Create an Assay object}
\usage{
CreateAssayObject(counts, var.names, key, rep, type, ...)

CreateAssayGrowthObject(counts, var.names, key, rep, na.rmax = 0.5)
}
\arguments{
\item{counts}{Unnormalized data such as raw counts}

\item{var.names}{A vector of variant names}

\item{key}{A character string for the experiment}

\item{rep}{Identity of the replicate}

\item{type}{Type of experiment: growth}

\item{...}{Additional arguments for create assay object.}

\item{na.rmax}{Maximum ratio of NA values allowed for a variant
to be included in the assay (AssayGrowth)}
}
\value{
An Assay object

An AssayGrowth object
}
\description{
Create an Assay object from a raw count matrix.

The expected format of the input matrix is variants x time points.
}
\details{
Non-unique variant names are not allowed. Please make unique before
calling this function. By default, variants with more than 0.5 missing (NA)
values will be filtered out. You can change the percentage by specifying
\code{na.rmax} or call \code{FilterData} after creating the rosace object.
}
