% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{CreateAssaySetObject}
\alias{CreateAssaySetObject}
\alias{CreateAssaySetGrowthObject}
\title{Create an AssaySet object}
\usage{
CreateAssaySetObject(combined.counts, var.names, reps, key, type, ...)

CreateAssaySetGrowthObject(
  combined.counts,
  var.names,
  reps,
  key,
  raw.counts,
  rounds
)
}
\arguments{
\item{combined.counts}{A matrix of counts (variants x time points in all reps)}

\item{var.names}{A vector of variant names}

\item{reps}{A vector of replicate identity}

\item{key}{A character string for the experiment}

\item{type}{Type of experiment: growth}

\item{...}{Additional arguments for create AssaySet object}

\item{raw.counts}{raw.counts A matrix of raw counts (variants x cols in reps)}

\item{rounds}{A vector of number of rounds}
}
\value{
An AssaySet object

An AssaySetGrowth object
}
\description{
Create an AssaySet object

Create an AssaySetGrowth object
}
