% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simconfig.R
\name{CreateConfig}
\alias{CreateConfig}
\title{Create the config list for simulation from a Rosette object}
\usage{
CreateConfig(
  object,
  n.sim,
  save.sim,
  type.sim,
  n.rep,
  null.var.group,
  wt.effect,
  n.round,
  n.mut,
  n.pos,
  pop.size,
  mode.sim = "clean",
  mode.rep = "bio",
  seq.shrink = 2,
  seq.depth = 200,
  lib.shrink = 2,
  var.shrink = 0.8,
  pos.flag = TRUE,
  ...
)
}
\arguments{
\item{object}{A Rosette object}

\item{n.sim}{Number of simulations}

\item{save.sim}{Directory to save the simulation data}

\item{type.sim}{Type of simulation: growth or binding}

\item{n.rep}{Number of replicates}

\item{null.var.group}{Name of the null variant group}

\item{wt.effect}{Wild-type effect (binding) or doubling rate (growth)}

\item{n.round}{Number of rounds}

\item{n.mut}{Number of mutants (currently has to be the same as in the Rosette object)
optional, can be inferred from the Rosette object.}

\item{n.pos}{Number of positions
optional, can be inferred from the Rosette object.}

\item{pop.size}{Cell population size
optional, by default 1000 cells per variant.}

\item{mode.sim}{Simulation mode: "clean" (default) or "reperror"}

\item{mode.rep}{Replicate mode: "bio" (default) or "tech"}

\item{seq.shrink}{Shrinkage factor for sequencing dispersion (times seq.disp)}

\item{seq.depth}{Sequencing depth, by default 200}

\item{lib.shrink}{Shrinkage factor for library dispersion}

\item{var.shrink}{Shrinkage factor for variance (var.dist)}

\item{pos.flag}{whether the simulation is position-centric}

\item{...}{Other parameters for noise.
Reperror: noise.freq, noise.mult}
}
\value{
A list of config
}
\description{
Create the config list for simulation from a Rosette object
}
