% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/rosace.R
\name{CreateRosaceObject}
\alias{CreateRosaceObject}
\alias{CreateRosaceObject.matrix}
\alias{CreateRosaceObject.Assay}
\title{Create a Rosace object}
\usage{
CreateRosaceObject(object, var.data, project.name = "RosaceProject", ...)

\method{CreateRosaceObject}{matrix}(object, var.data, project.name = "RosaceProject", type, ...)

\method{CreateRosaceObject}{Assay}(object, var.data, project.name = "RosaceProject", ...)
}
\arguments{
\item{object}{Either a matrix-like object with
unnormalized raw count or an Assay object}

\item{var.data}{Variants's meta data. If empty,
create a data frame with variants' name only.}

\item{project.name}{Name of the Rosace object}

\item{...}{Arguments passed to CreateAssayObject}

\item{type}{Type of assay to create: "growth"}
}
\value{
A Rosace object
}
\description{
Create a Rosace object from raw data
}
