% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/rosette.R
\name{CreateRosetteObject}
\alias{CreateRosetteObject}
\alias{CreateRosetteObject.Rosace}
\alias{CreateRosetteObject.Score}
\title{Create a Rosette object}
\usage{
CreateRosetteObject(object, project.name, ...)

\method{CreateRosetteObject}{Rosace}(
  object,
  project.name,
  score.name,
  pos.col,
  mut.col,
  ctrl.col,
  ctrl.name,
  ...
)

\method{CreateRosetteObject}{Score}(
  object,
  project.name,
  pos.label,
  mut.label,
  ctrl.label,
  disp,
  disp.start,
  ...
)
}
\arguments{
\item{object}{An object: numeric, Score, or Rosace}

\item{project.name}{Name of the Rosette object}

\item{...}{Additional arguments to be passed to the creation method.}

\item{score.name}{A character string indicating the name of the score}

\item{pos.col}{A character string indicating the name of the position column in var.data}

\item{mut.col}{A character string indicating the name of the mutant column in var.data}

\item{ctrl.col}{A character string indicating the name of the control column in var.data}

\item{ctrl.name}{A character string indicating the label of the control in the control column}

\item{pos.label}{A vector of position labels}

\item{mut.label}{A vector of mutant labels}

\item{ctrl.label}{A vector of control labels (boolean: ctrl or not)}

\item{disp}{Dispersion of raw count}

\item{disp.start}{Dispersion of variant library}
}
\value{
A Rosette object
}
\description{
Create a Rosette object
}
