% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{CreateScoreObject}
\alias{CreateScoreObject}
\title{Create a Score object}
\usage{
CreateScoreObject(method, type, assay.name, score, optional.score, misc)
}
\arguments{
\item{method}{A character string of the method used to calculate the score: SLR, ROSACE, ENRICH2}

\item{type}{A character string of the type: Assay, Assays}

\item{assay.name}{A character string of the Assay/Assays name}

\item{score}{A data frame of the score: The first three columns are variants, score, test-statistics.}

\item{optional.score}{A data frame of the optional score: The first column is variants.}

\item{misc}{A list of other information}
}
\description{
Create a Score object
}
