% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/simsummary.R
\name{EstimateDisp}
\alias{EstimateDisp}
\alias{EstimateDisp.AssayGrowth}
\alias{EstimateDisp.Rosace}
\title{Estimate dispersion from raw count (use all time points)}
\usage{
EstimateDisp(object, ...)

\method{EstimateDisp}{AssayGrowth}(object, ctrl.label, ...)

\method{EstimateDisp}{Rosace}(object, name, ctrl.col, ctrl.name, ...)
}
\arguments{
\item{object}{An object: AssayGrowth, or Rosace}

\item{...}{Additional arguments to be passed to the estimation method.}

\item{ctrl.label}{Boolean vector indicating control variants}

\item{name}{Name of AssayGrowth}

\item{ctrl.col}{Name of column in score.df that contains type of variants}

\item{ctrl.name}{Label of control variants in ctrl.col}
}
\description{
Estimate dispersion from raw count (use all time points)
}
