% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/simsummary.R
\name{EstimateDispStart}
\alias{EstimateDispStart}
\alias{EstimateDispStart.AssayGrowth}
\alias{EstimateDispStart.Rosace}
\title{Estimate dispersion from count0}
\usage{
EstimateDispStart(object, ...)

\method{EstimateDispStart}{AssayGrowth}(object, ...)

\method{EstimateDispStart}{Rosace}(object, name, ...)
}
\arguments{
\item{object}{An object: AssayGrowth, or Rosace}

\item{...}{Additional arguments to be passed to the estimation method.}

\item{name}{Name of AssayGrowth}
}
\description{
Estimate dispersion from count0
}
