% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{ExtractVarNames}
\alias{ExtractVarNames}
\title{Extract variants' names from a Score object}
\usage{
ExtractVarNames(score)
}
\arguments{
\item{score}{A Score object}
}
\value{
A character vector of variants' names
}
\description{
Extract variants' names from a Score object
}
