% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{FilterData}
\alias{FilterData}
\alias{FilterData.AssayGrowth}
\alias{FilterData.Rosace}
\title{Filter out variants in raw count data}
\usage{
FilterData(object, ...)

\method{FilterData}{AssayGrowth}(object, na.rmax = 0.5, min.count = 20, ...)

\method{FilterData}{Rosace}(object, name, key, ...)
}
\arguments{
\item{object}{An object: AssayGrowth (na.rm, min.count), or Rosace}

\item{...}{Additional arguments to be passed to the filtering method.}

\item{na.rmax}{The maximum ratio of NA values allowed in a variant}

\item{min.count}{The minimum number of counts needed for a variant}

\item{name}{character string of AssayGrowth name}

\item{key}{key of AssayGrowth}
}
\value{
An object with filtered count
}
\description{
Normalize the data by either a list of wild-type variants
or by the total count at the time point.
}
\details{
If normalizing by wild-type variants, the wild-type
variant names must be specified using \code{wt.var.names}
}
