% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosette.R
\name{GenerateOutput}
\alias{GenerateOutput}
\title{Generate summary statistics from a Rosette object}
\usage{
GenerateOutput(object, save.files = TRUE, save.dir)
}
\arguments{
\item{object}{A Rosette object}

\item{save.files}{A logical value indicating whether to save the output files}

\item{save.dir}{A directory to save the output files}
}
\description{
Including var.dist, mut.var.alpha, and mut.perc.
}
