% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsummary.R
\name{GenMutLabel}
\alias{GenMutLabel}
\title{Generate mutant group label in Rosette object}
\usage{
GenMutLabel(object, Gmut, hclust, save.plot = FALSE, save.dir = NULL)
}
\arguments{
\item{object}{An Rosette object}

\item{Gmut}{Number of mutant groups}

\item{hclust}{Input from function HclustMutant. If not given, will call HclustMutant.}

\item{save.plot}{If TRUE, save the plot to the directory save.dir}

\item{save.dir}{The directory to save the plot}
}
\value{
An Rosette object with mut.label and Gmut
}
\description{
Generate mutant group label in Rosette object
}
