% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/runROSACE.R
\name{GenRosaceInput}
\alias{GenRosaceInput}
\alias{GenRosaceInput.AssayGrowth}
\alias{GenRosaceInput.AssaySetGrowth}
\title{Generate input for Rosace model from Assay or AssaySet}
\usage{
GenRosaceInput(object, save.input, ...)

\method{GenRosaceInput}{AssayGrowth}(
  object,
  save.input,
  pos.label,
  ctrl.label,
  stop.label,
  thred = 10,
  ...
)

\method{GenRosaceInput}{AssaySetGrowth}(
  object,
  save.input,
  pos.label,
  ctrl.label,
  stop.label,
  thred = 10,
  ...
)
}
\arguments{
\item{object}{Assay or AssaySet}

\item{save.input}{if given save input to the save.input path}

\item{...}{Additional arguments to be passed to the input generation method.}

\item{pos.label}{vector of true position of variants}

\item{ctrl.label}{vector of whether variant is control}

\item{stop.label}{vector of whether variant is stop/nonsense}

\item{thred}{integer, threshold for number of variants per position label (index)
Passed to function "varPosIndexMap"}
}
\value{
A list of input for Rosace1 model
}
\description{
Generate input for Rosace model from Assay or AssaySet
}
