% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsummary.R
\name{GenVarLabel}
\alias{GenVarLabel}
\title{Generate variant group label in Rosette object}
\usage{
GenVarLabel(object, Gvar)
}
\arguments{
\item{object}{An Rosette object}

\item{Gvar}{Number of variant groups}
}
\value{
An Rosette object with var.dist and Gvar
}
\description{
Generate variant group label in Rosette object
}
