% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosette.R
\name{GetScoreWide}
\alias{GetScoreWide}
\title{Return wide dataframe of score in a Rosette object}
\usage{
GetScoreWide(object)
}
\arguments{
\item{object}{A Rosette object}
}
\value{
A "wide" data.frame
}
\description{
Row is position and column is mutant. The entry is scores.
}
