% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsummary.R
\name{HclustMutant}
\alias{HclustMutant}
\title{Hierarchical clustering the mutant in Rosette object}
\usage{
HclustMutant(object, save.plot = FALSE, save.dir = NULL)
}
\arguments{
\item{object}{An Rosette object}

\item{save.plot}{If TRUE, save the plot to the directory save.dir}

\item{save.dir}{The directory to save the plot}
}
\description{
Determine the number of mutant group from the plot.
}
