% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSACE.R
\name{imputeAssaysCount}
\alias{imputeAssaysCount}
\title{Impute Assays count}
\usage{
imputeAssaysCount(counts, rounds)
}
\arguments{
\item{counts}{A matrix of counts}

\item{rounds}{A vector of number of rounds}
}
\value{
A list of imputed counts and rounds
}
\description{
If replicates are with different number of rounds,
truncate counts beyond the miminum number of rounds.
Otherwise, impute with mean counts of replicates.
}
