% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSACE.R
\name{imputeAssaysCountKNN}
\alias{imputeAssaysCountKNN}
\title{Impute Assays count (KNN Package)}
\usage{
imputeAssaysCountKNN(counts, rounds)
}
\arguments{
\item{counts}{A matrix of counts}

\item{rounds}{A vector of number of rounds}
}
\value{
A list of imputed counts and rounds
}
\description{
Impute Assays count (KNN Package)
}
