% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{IntegrateData}
\alias{IntegrateData}
\alias{IntegrateData.Assay}
\alias{IntegrateData.AssaySet}
\alias{IntegrateData.Rosace}
\title{Integrate Assay into Assays}
\usage{
IntegrateData(object, ...)

\method{IntegrateData}{Assay}(object, assay, ...)

\method{IntegrateData}{AssaySet}(object, assay, ...)

\method{IntegrateData}{Rosace}(object, key, ...)
}
\arguments{
\item{object}{An object: Assay, AssaySet, or Rosace}

\item{...}{Additional arguments to be passed to the integration method.}

\item{assay}{An object from class 'Assay'}

\item{key}{The key for the experiment in the AssaySet object}
}
\value{
An object with integrated count (AssaySet)
}
\description{
Integrate Assay into Assays
}
