% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/runROSACE.R
\name{MCMCCreateScore}
\alias{MCMCCreateScore}
\alias{MCMCCreateScore.Assay}
\alias{MCMCCreateScore.AssaySet}
\title{Create score object from MCMC result}
\usage{
MCMCCreateScore(object, main.score, param.post, diags)

\method{MCMCCreateScore}{Assay}(object, main.score, param.post, diags)

\method{MCMCCreateScore}{AssaySet}(object, main.score, param.post, diags)
}
\arguments{
\item{object}{Assay or AssaySet}

\item{main.score}{functional score}

\item{param.post}{(optional) posterior of parameters other than "functional score"}

\item{diags}{(optional) diagnostics of MCMC}
}
\value{
A Score object
}
\description{
Create score object from MCMC result
}
