% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSACE.R
\name{MCMCLfsr}
\alias{MCMCLfsr}
\title{Compute lfsr for variables from CmdStanMCMC object}
\usage{
MCMCLfsr(fit, param.key = "beta")
}
\arguments{
\item{fit}{CmdStanMCMC object}

\item{param.key}{character string to match parameter names (functinal score)}
}
\value{
numeric vector of lfsr for each parameter
}
\description{
Compute lfsr for variables from CmdStanMCMC object
}
