% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSACE.R
\name{MCMCRunStan}
\alias{MCMCRunStan}
\title{Run Stan MCMC with input list and compiled model}
\usage{
MCMCRunStan(input, mod, seed = 100, refresh = 100)
}
\arguments{
\item{input}{list of input data}

\item{mod}{compiled model}

\item{seed}{random seed}

\item{refresh}{number of iterations between progress updates}
}
\value{
CmdStanMCMC object
}
\description{
Run Stan MCMC with input list and compiled model
}
