% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSACE.R
\name{MCMCScoreDf}
\alias{MCMCScoreDf}
\title{Extract "functional score" posterior distribution from CmdStanMCMC object}
\usage{
MCMCScoreDf(fit, param.key, param.post, savefile, output.lfsr = TRUE)
}
\arguments{
\item{fit}{CmdStanMCMC object}

\item{param.key}{parameter key to extract}

\item{param.post}{posterior distribution of parameters (optional)
If not given, extract from fit object.}

\item{savefile}{If given, save tsv output to file}

\item{output.lfsr}{whether to compute lfsr}
}
\value{
data.frame of "functional score" posterior distribution
}
\description{
Extract "functional score" posterior distribution from CmdStanMCMC object
}
