% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R, R/score.R
\name{names.Assay}
\alias{names.Assay}
\alias{names.AssaySet}
\alias{names.Score}
\title{Return the name of the Assay object}
\usage{
\method{names}{Assay}(x)

\method{names}{AssaySet}(x)

\method{names}{Score}(x)
}
\arguments{
\item{x}{A Score object}
}
\value{
A character string of the name of the Assay object

A character string of the name of the AssaySetobject

A character string of the name of the Score object
}
\description{
Return the name of the Assay object

Return the name of the AssaySet object

Return the name of the Score object
}
