% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{NormalizeData}
\alias{NormalizeData}
\alias{NormalizeData.matrix}
\alias{NormalizeData.AssayGrowth}
\alias{NormalizeData.Rosace}
\title{Normalize count (AssayGrowth)}
\usage{
NormalizeData(object, normalization.method, ...)

\method{NormalizeData}{matrix}(object, normalization.method, var.names, ...)

\method{NormalizeData}{AssayGrowth}(object, normalization.method, ...)

\method{NormalizeData}{Rosace}(object, normalization.method, name, key, ...)
}
\arguments{
\item{object}{An object: matrix, AssayGrowth, or Rosace}

\item{normalization.method}{The normalization method to use: "wt" or "total"}

\item{...}{Additional arguments to be passed to the normalization method.}

\item{var.names}{a vector of variant names to be normalized}

\item{name}{character string of AssayGrowth name}

\item{key}{key of AssayGrowth}
}
\value{
An object with normalized count
}
\description{
Normalize (filtered and imputed) count data (in Assay)
and get normalized count
}
