% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oct1}
\alias{oct1}
\alias{oct1_rep1}
\alias{oct1_rep2}
\alias{oct1_rep3}
\title{OCT1 drug cytotoxicity screen: raw count}
\format{
A data frame with variant as rows. The first column is the variant
label, and the rest of them are raw counts at different time points in order.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11440 rows and 5 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11453 rows and 5 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11424 rows and 4 columns.
}
\usage{
data(oct1)

oct1_rep1

oct1_rep2

oct1_rep3
}
\description{
A dataset containing three replicates of OCT1 drug cytotoxicity screen.
Paper link: https://www.biorxiv.org/content/10.1101/2023.06.06.543963v1.full
}
\keyword{datasets}
