% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oct1_rosace}
\alias{oct1_rosace}
\title{OCT1 drug cytotoxicity screen: processed rosace object}
\format{
A rosace object with three assays (replicates) and one assayset.
}
\usage{
data(oct1_rosace)
}
\description{
OCT1 drug cytotoxicity screen: processed rosace object
}
\keyword{datasets}
