% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oct1_rosace_scored}
\alias{oct1_rosace_scored}
\title{OCT1 drug cytotoxicity screen: effect score computed rosace object}
\format{
A rosace object with three assays (replicates), one assayset, and one score.
}
\usage{
data(oct1_rosace_scored)
}
\description{
OCT1 drug cytotoxicity screen: effect score computed rosace object
}
\keyword{datasets}
