% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/simsummary.R
\name{PlotDisp}
\alias{PlotDisp}
\alias{PlotDisp.AssayGrowth}
\alias{PlotDisp.Rosace}
\title{Raw count dispersion plot}
\usage{
PlotDisp(object, t, ...)

\method{PlotDisp}{AssayGrowth}(object, t, ctrl.label, ...)

\method{PlotDisp}{Rosace}(object, t, name, ctrl.col, ctrl.name, ...)
}
\arguments{
\item{object}{An object: matrix, Assay, or Rosace}

\item{t}{The time point to compute dispersion (>= 2). If missing, use all time points.}

\item{...}{Additional arguments to be passed to the plot method.}

\item{ctrl.label}{Boolean vector indicating control variants}

\item{name}{Name of AssayGrowth}

\item{ctrl.col}{Name of column in score.df that contains type of variants}

\item{ctrl.name}{Label of control variants in type.col}
}
\description{
Raw count dispersion plot
}
