% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsummary.R
\name{PMVCountDist}
\alias{PMVCountDist}
\title{Infer distribution for number of variants for each variant and
mutant group label at each position}
\usage{
PMVCountDist(object, pos.missing = 0.2)
}
\arguments{
\item{object}{A Rosette object}

\item{pos.missing}{Percentage of variants missing at each position allowed}
}
\value{
A Rosette object with mut.var.alpha
}
\description{
Fit a Dirichlet distribution for number of variants in
(mutant group, variant group) at each position
}
