% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simconfig.R
\name{ReadFullConfig}
\alias{ReadFullConfig}
\alias{ReadPathConfig}
\title{Read a full config file
(with null.var.group, mut.var.alpha, var.dist as data frame)}
\usage{
ReadFullConfig(file)

ReadPathConfig(file)
}
\arguments{
\item{file}{Config file path}
}
\value{
A list of config

A list of config
}
\description{
Read a full config file
(with null.var.group, mut.var.alpha, var.dist as data frame)

Read a path config file
(with null.var.group, mut.var.alpha, var.dist as file path)
}
