% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosace.R
\docType{class}
\name{Rosace-class}
\alias{Rosace-class}
\alias{Rosace}
\title{The Rosace Class}
\description{
The Rosace Class
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of Assay object}

\item{\code{assay.sets}}{A list of AssaySet object}

\item{\code{var.data}}{Contains meta-information about each variant.
First column has to contain variants' names}

\item{\code{scores}}{A list of Score object}

\item{\code{project.name}}{Name of the project}

\item{\code{misc}}{A list of other information}
}}

