% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosace.R
\name{ExtractAssay}
\alias{ExtractAssay}
\alias{ExtractAssaySet}
\alias{ExtractScore}
\title{Extract Specific Object Type}
\usage{
ExtractAssay(object, name)

ExtractAssaySet(object, name)

ExtractScore(object, name)
}
\arguments{
\item{object}{A Rosace object}

\item{name}{The targeted name of Assay/AssaySet/Score object}
}
\value{
The Assay/AssaySet/Score object
}
\description{
Get the names of Assay, AssaySet, Score objects
}
