% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosace.R
\name{GetAssayName}
\alias{GetAssayName}
\alias{GetAssaySetName}
\alias{GetScoreName}
\title{Get the Names of Specific Object Type}
\usage{
GetAssayName(object)

GetAssaySetName(object)

GetScoreName(object)
}
\arguments{
\item{object}{A Rosace object}
}
\value{
The names of all component objects in this Rosace object
}
\description{
Get the names of Assay, AssaySet, Score objects
}
