% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosette.R
\docType{class}
\name{Rosette-class}
\alias{Rosette-class}
\alias{Rosette}
\title{The Rosette Class}
\description{
A Rosette object is a container for inferring summary statistics from a
AssayGrowth of a Rosace experiment.
}
\section{Slots}{

\describe{
\item{\code{score.df}}{A data.frame with columns score, pos, mut, ctrl, (var_group),
(mut_group)}

\item{\code{mut.label}}{A data.frame with columns mut, mut_group}

\item{\code{var.dist}}{A data.frame with columns var_group, mean, sd, count}

\item{\code{mut.var.alpha}}{A data.frame with columns mut_group, var_group, alpha,
count}

\item{\code{disp}}{Dispersion of raw count (sequencing step)}

\item{\code{disp.start}}{Dispersion of starting variant library}

\item{\code{Gvar}}{Number of variant groups}

\item{\code{Gmut}}{Number of mutant groups}

\item{\code{rounds}}{Number of rounds}

\item{\code{project.name}}{A character string indicating the name of the project}
}}

