% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSETTE.R
\name{runRosette}
\alias{runRosette}
\title{Simulate data from a config file}
\usage{
runRosette(config, save.tsv = TRUE, save.rosace = TRUE, save.enrich2 = FALSE)
}
\arguments{
\item{config}{A list of config}

\item{save.tsv}{A boolean indicating whether to save the data in tsv format}

\item{save.rosace}{A boolean indicating whether to save the Rosace object}

\item{save.enrich2}{A boolean indicating whether to save the data for enrich2}
}
\value{
None
}
\description{
Simulate data from a config file
}
