% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/runBASE.R
\name{runSLR}
\alias{runSLR}
\alias{runSLR.AssayGrowth}
\alias{runSLR.AssaySetGrowth}
\alias{runSLR.Rosace}
\title{Do simple linear regression on normalized count  (Growth)}
\usage{
runSLR(object, ...)

\method{runSLR}{AssayGrowth}(object, ...)

\method{runSLR}{AssaySetGrowth}(object, ...)

\method{runSLR}{Rosace}(object, name, type, ...)
}
\arguments{
\item{object}{An object: AssayGrowth, AssaySetGrowth, or Rosace}

\item{...}{Additional arguments to be passed to the regression method.}

\item{name}{character string of Assay/AssaySet name}

\item{type}{Assay or AssaySet}
}
\value{
A Score object with regression result
}
\description{
For each variant, fit a simple linear regression on all replicates across
different time points. The regression coefficient, standard error, p-value
and any optional scores are returned as part of a Score object.
}
