% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\docType{class}
\name{Score-class}
\alias{Score-class}
\alias{Score}
\title{The Score Class}
\description{
The Score Class
}
\section{Slots}{

\describe{
\item{\code{method}}{A character string of the method used to calculate the score}

\item{\code{type}}{A character string of the type: Assay, AssaySet}

\item{\code{assay.name}}{A character string of the Assay/AssaySet name}

\item{\code{score}}{A data frame of the score:
The first four columns are variants, score, sd, test-statistics.}

\item{\code{optional.score}}{A data frame of the optional score
(row corresponding to score)}

\item{\code{misc}}{A list of other information}
}}

