% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scoreDensity}
\alias{scoreDensity}
\title{Generate Score Density Plot}
\usage{
scoreDensity(
  data,
  savedir,
  type.col = "type",
  score.col = "mean",
  hist = FALSE,
  nbins = 30,
  c.fill = c("#FF7575", "lightgreen", "#7298BF", "#E0B0FF", "#FFDAB9", "#40E0D0"),
  alpha = 0.5,
  x.text = 10,
  y.text = 10,
  scale.free = FALSE,
  space.free = FALSE,
  ht = 10,
  wd = 8,
  name = "DensityPlot",
  savepdf = TRUE,
  show = TRUE
)
}
\arguments{
\item{data}{Scores data frame. Expected to have columns containing information about
position, control amino acid, mutated amino acid, mutation type, and score.}

\item{savedir}{Character string specifying the directory to save plots.}

\item{type.col}{Column name in \code{data} for mutation types. Default is \code{type}.}

\item{score.col}{Column name in \code{data} for mutation scores. Default is \code{mean}.}

\item{hist}{Logical indicating whether to plot count histogram or density.
Default is \code{FALSE}.}

\item{nbins}{Numeric value specifying the number of bins for the histogram.
Default is \code{30}.}

\item{c.fill}{Vector indicating the fill color for mutation types.}

\item{alpha}{Numeric value between 0-1 indicating the fill transparency.
Default is \code{0.5}}

\item{x.text}{Numeric value for x-axis text size. Default is \code{10}.}

\item{y.text}{Numeric value for x-axis text size. Default is \code{10}.}

\item{scale.free}{Logical indicating whether to make score range proportional.
Default is \code{FALSE}.}

\item{space.free}{Logical indicating whether to make panel heights variable.
Default is \code{FALSE}.}

\item{ht}{Numeric value for the height of the saved plot. Default is \code{10}.}

\item{wd}{Numeric value for the width of the saved plot. Default is \code{8}.}

\item{name}{Character string specifying the base name of the saved file.}

\item{savepdf}{Logical indicating whether to also save a PDF version of the plot.
Default is \code{TRUE}.}

\item{show}{Logical indicating whether or not to display the plot in the viewer.
Default is \code{TRUE}.}
}
\value{
NULL.
}
\description{
\code{scoreDensity} generates a density plot for visualizing the distribution of scores
across different mutation types.
}
\examples{
\dontrun{
scoreDensity(data = scores.data, scale.free = TRUE,
            savedir = "./plots/", name = "DensityPlot")
}

}
