% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scoreHeatmap}
\alias{scoreHeatmap}
\title{Generate Score Heatmap}
\usage{
scoreHeatmap(
  data,
  savedir,
  ctrl.name = "synonymous",
  pos.col = "position",
  wt.col = "wildtype",
  mut.col = "mutation",
  type.col = "type",
  score.col = "mean",
  aa.order = NA,
  npos = 100,
  ncol = 1,
  pos.step = 5,
  x.text = 6,
  y.text = 3,
  seq.text = 1.1,
  c.pallete = "RdBu",
  c.scale = list(),
  ht = 11,
  wd = 8.5,
  name = "Heatmap",
  savepdf = TRUE,
  show = FALSE
)
}
\arguments{
\item{data}{Scores data frame. Expected to have columns containing information about
position, control amino acid, mutated amino acid, mutation type, and score.}

\item{savedir}{Character string specifying the directory to save plots.}

\item{ctrl.name}{String specifying the control mutation name. Default is \code{synonymous}.}

\item{pos.col}{Column name in \code{data} for mutation positions. Default is \code{position}.}

\item{wt.col}{Column name in \code{data} for wildtype amino acids. Default is \code{wildtype}.}

\item{mut.col}{Column name in \code{data} for mutated amino acids. Default is \code{mutation}.}

\item{type.col}{Column name in \code{data} for mutation types. Default is \code{type}.}

\item{score.col}{Column name in \code{data} for mutation scores. Default is \code{mean}.}

\item{aa.order}{Character vector defining the order of amino acid mutations in the y-axis.
Default to using mutations from \code{data} in alphabetical order.}

\item{npos}{Integer specifying the number of positions per subplot. Default is \code{100}.}

\item{ncol}{Integer specifying the number of columns of subplots. Default is \code{1}.}

\item{pos.step}{Integer specifying the steps between x-axis labels. Default is \code{5}.}

\item{x.text}{Numeric value for x-axis text size. Default is \code{6}.}

\item{y.text}{Numeric value for y-axis text size. Default is \code{3}.}

\item{seq.text}{Numeric value for wildtype sequence text size. Default is \code{1.1}}

\item{c.pallete}{Character string or vector defining the color palette. Default is \code{'RdBu'}.}

\item{c.scale}{List of parameters definning the score color scale.}

\item{ht}{Numeric value for the height of the saved plot. Default is \code{11}.}

\item{wd}{Numeric value for the width of the saved plot. Default is \code{8.5}.}

\item{name}{Character string specifying the base name of the saved file.}

\item{savepdf}{Logical indicating whether to also save a PDF version of the plot.
Default is \code{TRUE}.}

\item{show}{Logical indicating whether or not to display the plot in the viewer.
Default is \code{FALSE}.}
}
\value{
NULL.
}
\description{
\code{scoreHeatmap} generates a series of heatmaps for visualizing the
score patterns across different positions.
}
\examples{
\dontrun{
scoreHeatmap(scores.data, savedir = "./plots/", name = "Heatmap", savepdf = TRUE)
}

}
