% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scoreVlnplot}
\alias{scoreVlnplot}
\title{Generate Score Violin Plot}
\usage{
scoreVlnplot(
  data,
  savedir,
  pos.col = "position",
  wt.col = "wildtype",
  score.col = "mean",
  jitter = FALSE,
  c.fill = "lightblue",
  npos = 50,
  ncol = 1,
  pos.step = 1,
  x.text = 8,
  y.text = 8,
  seq.text = 3,
  pt.size = 0.2,
  ht = 30,
  wd = 15,
  name = "Violinplot",
  savepdf = TRUE,
  show = FALSE
)
}
\arguments{
\item{data}{Scores data frame. Expected to have columns containing information about
position, control amino acid, mutated amino acid, mutation type, and score.}

\item{savedir}{A string providing the directory path to save plots.}

\item{pos.col}{Column name in \code{data} for mutation positions. Default is \code{position}.}

\item{wt.col}{Column name in \code{data} for wildtype amino acids. Default is \code{wildtype}.}

\item{score.col}{Column name in \code{data} for mutation scores. Default is \code{mean}.}

\item{jitter}{Logical indicating whether to add jitter points to violin plot.
Default is \code{FALSE}.}

\item{c.fill}{String indicating the fill color for the violin plots.}

\item{npos}{Integer specifying the number of positions per subplot. Default is \code{50}.}

\item{ncol}{Integer specifying the number of columns of subplots. Default is \code{1}.}

\item{pos.step}{Integer specifying the steps between x-axis labels. Default is \code{5}.}

\item{x.text}{Numeric value for x-axis text size. Default is \code{8}.}

\item{y.text}{Numeric value for x-axis text size. Default is \code{8}.}

\item{seq.text}{Numeric value for wildtype sequence text size. Default is \code{3}}

\item{pt.size}{Numeric value specifying the size of jitter points. Default is \code{0.2}}

\item{ht}{Numeric value for the height of the saved plot. Default is \code{30}.}

\item{wd}{Numeric value for the width of the saved plot. Default is \code{15}.}

\item{name}{Character string specifying the base name of the saved file.}

\item{savepdf}{Logical indicating whether to also save a PDF version of the plot.
Default is \code{TRUE}.}

\item{show}{Logical indicating whether or not to display the plot in the viewer.
Default is \code{FALSE}.}
}
\value{
NULL.
}
\description{
\code{scoreVlnplot} generates a series of violin plots for visualizing the
distribution of scores across different positions.
}
\examples{
\dontrun{
scoreVlnplot(data = scores.data, jitter = TRUE, c.fill = "lightpink",
            savedir = "./plots/", name = "ViolinPlot")
}

}
