% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBASE.R
\name{SLR}
\alias{SLR}
\title{Internal function for simple linear regression}
\usage{
SLR(y, x)
}
\arguments{
\item{y}{numeric vector of response}

\item{x}{numeric vector of predictor}
}
\value{
numeric vector of regression coefficients
}
\description{
Internal function for simple linear regression
}
