% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosace.R
\name{ExtractVarAssay}
\alias{ExtractVarAssay}
\alias{ExtractVarAssaySet}
\alias{ExtractVarScore}
\title{Extract the Meta Data for Variants in the Specific Object Type}
\usage{
ExtractVarAssay(object, name, norm = TRUE)

ExtractVarAssaySet(object, name)

ExtractVarScore(object, name)
}
\arguments{
\item{object}{A Rosace object}

\item{name}{The targeted name of Assay/AssaySet/Score object}

\item{norm}{(for AssayGrowth only) use the 'var.names' or 'norm.var.names'}
}
\value{
Meta data for variants in Assay, AssaySet, Score object
}
\description{
Extract the meta data for variants in Assay, AssaySet, Score object
}
