% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSACE.R
\name{varPosIndexMap}
\alias{varPosIndexMap}
\title{Map each variant to its position label (index)}
\usage{
varPosIndexMap(var.names, pos.label, ctrl.label, stop.label, thred = 10)
}
\arguments{
\item{var.names}{character vector of variant names}

\item{pos.label}{character or numeric vector of true positions}

\item{ctrl.label}{vector of whether variant is control}

\item{stop.label}{vector of whether variant is stop/nonsense}

\item{thred}{integer, threshold for number of variants per position label (index)}
}
\value{
data.frame with columns 'variant', 'pos', 'index'
}
\description{
Threshold is tricky to choose.
}
