% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simconfig.R
\name{WriteConfig}
\alias{WriteConfig}
\title{Write the config list in json format}
\usage{
WriteConfig(cfg, file, savedir)
}
\arguments{
\item{cfg}{A list of config}

\item{file}{File name}

\item{savedir}{Directory to save the file}
}
\value{
None
}
\description{
Write the config list in json format
}
