% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runROSACE.R
\name{CheckStanSetup}
\alias{CheckStanSetup}
\title{Check Stan Setup}
\usage{
CheckStanSetup(mc.cores, install.update = TRUE)
}
\arguments{
\item{mc.cores}{Number of cores to use for parallel builds}

\item{install.update}{whether to update CmdStan}
}
\value{
None
}
\description{
Check Stan Setup
}
