% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocessing.R
\name{FilterData}
\alias{FilterData}
\alias{FilterData.AssayGrowth}
\alias{FilterData.Rosace}
\title{Filter out variants in raw count data}
\usage{
FilterData(object, ...)

\method{FilterData}{AssayGrowth}(object, na.rmax = 0.5, ...)

\method{FilterData}{Rosace}(object, name, key, ...)
}
\arguments{
\item{object}{An object: AssayGrowth (na.rm), or Rosace}

\item{...}{Additional arguments to be passed to the filtering method.}

\item{na.rmax}{The maximum ratio of NA values allowed in a variant}

\item{name}{character string of AssayGrowth name}

\item{key}{key of AssayGrowth}
}
\value{
An object with filtered count
}
\description{
Filter out variants in raw count data
}
