% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/score.R
\name{OutputScore}
\alias{OutputScore}
\alias{OutputScore.Score}
\alias{OutputScore.Rosace}
\title{Output Score data frame}
\usage{
OutputScore(object, ...)

\method{OutputScore}{Score}(object, ...)

\method{OutputScore}{Rosace}(object, name, ...)
}
\arguments{
\item{object}{Either a Rosace object or a Score object}

\item{...}{Additional arguments passed to OutputScore}

\item{name}{The targeted name of Score object}
}
\value{
A dataframe of score
}
\description{
If the object is a Rosace object, output the meta-information
of the variants in the data frame as well.
}
