% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/runROSACE.R
\name{RunRosace}
\alias{RunRosace}
\alias{RunRosace.AssayGrowth}
\alias{RunRosace.AssaySetGrowth}
\alias{RunRosace.Rosace}
\title{Run Rosace on an Assay/AssaySet object}
\usage{
RunRosace(object, savedir, mc.cores, debug, install, ...)

\method{RunRosace}{AssayGrowth}(
  object,
  savedir,
  mc.cores = 4,
  debug = FALSE,
  install = TRUE,
  pos.label,
  ctrl.label,
  ...
)

\method{RunRosace}{AssaySetGrowth}(
  object,
  savedir,
  mc.cores = 4,
  debug = FALSE,
  install = TRUE,
  pos.label,
  ctrl.label,
  ...
)

\method{RunRosace}{Rosace}(
  object,
  savedir,
  mc.cores = 4,
  debug = FALSE,
  install = TRUE,
  name,
  type,
  pos.col,
  ctrl.col,
  ctrl.name,
  ...
)
}
\arguments{
\item{object}{Rosace, Assay/AssaySet (default)}

\item{savedir}{directory to save the output}

\item{mc.cores}{integer, number of cores to use for parallel computing}

\item{debug}{logical, if TRUE, return a list of cmdstanfit and Score object.}

\item{install}{logical, if TRUE, install or update cmdstanr
There's a chance for Github download error if running multiple instances on a server.}

\item{...}{Additional arguments to be passed to the run method.}

\item{pos.label}{vector of true position of variants}

\item{ctrl.label}{vector of whether variant is in the control group (NA if none provided)}

\item{name}{Name of the object to be analyzed}

\item{type}{"Assay" or "AssaySet"}

\item{pos.col}{For Growth screen, the column name for position in the var.data
(optional in no_pos mode)}

\item{ctrl.col}{For Growth screen, optional for control to have one position index}

\item{ctrl.name}{For Growth screen, optional, the name of the control type}
}
\value{
A object with Rosace result (Score object) if debug = FALSE
}
\description{
Run Rosace on an Assay/AssaySet object
}
